feat: change channel from command

+++ src/receiver2.h
@@ -24,7 +24,7 @@ void setupReceiverPins();
 void setupDefaultChannels();
 void sendRegister(uint8_t addressBits, uint32_t dataBits, int receiverPin);
 void readRssiFromReceivers();
-void changeChannel(uint16_t value, int receiverPin);
+void changeChannel(uint8_t value, int receiverPin);
 int getPinForReceiver(int receiverId);
 void startReadingRssi(int status, int64_t millis_start);
